DEV: Apply code suggestions
